Server/requirements.txt
Django==2.0.1 
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
django-allauth==0.34.0
django-rest-auth==0.9.2
django-cors-headers==2.1.0